Initial code review feedback
