--- src/Microsoft.Diagnostics.DebugServices.Implementation/MemoryRegionService.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-
-namespace Microsoft.Diagnostics.DebugServices.Implementation
-{
-    public class MemoryRegionService : IMemoryRegionService
-    {
-        public IEnumerable<IMemoryRegion> EnumerateRegions()
-        {
-            throw new NotImplementedException();
-        }
-    }
-}
